requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
black==26.1.0
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
filelock==3.24.0
greenlet==3.3.1
h11==0.16.0
identify==2.6.16
idna==3.11
librt==0.8.0
Mako==1.3.10
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.1
pre_commit==4.5.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.4.1
PyYAML==6.0.3
rsa==4.9.1
ruff==0.15.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
virtualenv==20.36.1
